2 - Deep Learning/ASL Arabic Tracking Web App – Quran Kareem/requirements.txt
streamlit==1.48.0
ultralytics==8.3.174
torch==2.1.1+cu118
torchvision==0.16.1+cu118
torchaudio==2.1.1+cu118
opencv-python==4.12.0.88
numpy==1.26.4
pillow==11.3.0
pandas==2.3.1
matplotlib==3.10.5
seaborn==0.13.2
scikit-learn==1.7.1
tqdm==4.67.1
python-bidi==0.6.6
arabic-reshaper==3.0.0
